Moving tests into /components folder.
