+++ __tests__/components/Header-test.tsx
@@ -2,7 +2,7 @@ import { expect } from "chai";
 
 import * as React from "react";
 import * as Enzyme from "enzyme";
-import Header from "../src/components/Header";
+import Header from "../../src/components/Header";
 
 describe("Header", () => {
   let wrapper: Enzyme.ShallowWrapper<any, {}>;
