+++ __tests__/components/Input-test.tsx
@@ -1,7 +1,7 @@
 import { expect } from "chai";
 import * as Enzyme from "enzyme";
 import * as React from "react";
-import Input from "../src/components/Input";
+import Input from "../../src/components/Input";
 
 describe("Input", () => {
   let wrapper: Enzyme.ShallowWrapper<any, {}>;
