+++ __tests__/components/Panel-test.tsx
@@ -3,7 +3,7 @@ import { expect } from "chai";
 import * as React from "react";
 import * as Enzyme from "enzyme";
 import * as Sinon from "sinon";
-import Panel from "../src/components/Panel";
+import Panel from "../../src/components/Panel";
 
 describe("Panel", () => {
   let wrapper: Enzyme.CommonWrapper<any, any, {}>;
